# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para plataforma de serviços de limpeza.",
        "title": "Phresh-Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/usuarios/eu/perfil": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Buscar perfil do usuário autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfil"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Atualizar perfil base do usuário autenticado",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoAtualizarPerfil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfil"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/usuarios/eu/perfil-cliente": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Buscar perfil de cliente do usuário autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilCliente"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Atualizar perfil de cliente do usuário autenticado",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoAtualizarPerfilCliente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilCliente"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Criar perfil de cliente para o usuário autenticado",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilCliente"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/usuarios/eu/perfil-faxineiro": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Buscar perfil de faxineiro do usuário autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilFaxineiro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Atualizar perfil de faxineiro do usuário autenticado",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoAtualizarPerfilFaxineiro"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilFaxineiro"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Criar perfil de faxineiro para o usuário autenticado",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaPerfilFaxineiro"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
        "/usuarios/registrar": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usuarios"
                ],
                "summary": "Registrar novo usuário",
                "parameters": [
                    {
                        "description": "Dados de registro",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoRegistro"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaUsuario"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.EnderecoDTO": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "complemento": {
                    "type": "string"
                },
                "estado": {
                    "type": "string"
                },
                "rua": {
                    "type": "string"
                }
            }
        },
...
        "dto.RequisicaoAtualizarPerfil": {
            "type": "object",
            "properties": {
                "imagem": {
                    "type": "string"
                },
                "nome_completo": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        },
        "dto.RequisicaoAtualizarPerfilCliente": {
            "type": "object",
            "properties": {
                "banheiros": {
                    "type": "integer"
                },
                "endereco": {
                    "$ref": "#/definitions/dto.EnderecoDTO"
                },
                "observacoes": {
                    "type": "string"
                },
                "quartos": {
                    "type": "integer"
                },
                "tamanho_imovel_m2": {
                    "type": "number"
                },
                "tipo_imovel": {
                    "type": "string"
                }
            }
        },
        "dto.RequisicaoAtualizarPerfilFaxineiro": {
            "type": "object",
            "properties": {
                "anos_experiencia": {
                    "type": "integer"
                },
                "cidades_atendidas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "descricao": {
                    "type": "string"
                },
                "especialidades": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.RequisicaoRegistro": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nome_usuario": {
                    "type": "string"
                }
            }
        },
...
        "dto.RespostaErro": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "detalhes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mensagem": {
                    "type": "string"
                }
            }
        },
...
        "dto.RespostaPerfil": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "imagem": {
                    "type": "string"
                },
                "nome_completo": {
                    "type": "string"
                },
                "nome_usuario": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "usuario_id": {
                    "type": "integer"
                }
            }
        },
        "dto.RespostaPerfilCliente": {
            "type": "object",
            "properties": {
                "banheiros": {
                    "type": "integer"
                },
                "endereco": {
                    "$ref": "#/definitions/dto.EnderecoDTO"
                },
                "faxineiro_preferido_id": {
                    "type": "integer"
                },
                "observacoes": {
                    "type": "string"
                },
                "quartos": {
                    "type": "integer"
                },
                "tamanho_imovel_m2": {
                    "type": "number"
                },
                "tipo_imovel": {
                    "type": "string"
                },
                "usuario_id": {
                    "type": "integer"
                }
            }
        },
        "dto.RespostaPerfilFaxineiro": {
            "type": "object",
            "properties": {
                "anos_experiencia": {
                    "type": "integer"
                },
                "cidades_atendidas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "descricao": {
                    "type": "string"
                },
                "especialidades": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "usuario_id": {
                    "type": "integer"
                },
                "verificado": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.RespostaUsuario": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome_usuario": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}